requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
torchvision>=0.16.0
torchaudio>=2.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Data processing
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Utils
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Model utilities
timm>=0.9.0  # PyTorch Image Models
transformers>=4.30.0  # Hugging Face transformers

# Experiment tracking (optional)
# wandb>=0.15.0
# mlflow>=2.5.0
